Time: 1 ms (58%), Space: 8.7 MB (38.09%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        long y=x;
+        int y=x;
         long rev = 0;
         while (x>0)
         {
